backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
python-dotenv
requests
tensorflow
numpy